Django Application/requirements.txt
altair==5.3.0
asgiref==3.8.1
astor==0.8.1
attrs==23.2.0
backcall==0.2.0
base58==2.1.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
chardet==5.2.0
click==8.1.7
cmake==3.29.6
colorama==0.4.6
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.6
dlib==19.24.2
docutils==0.21.2
entrypoints==0.4
enum-compat==0.0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
future==1.0.0
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.3
jedi==0.16.0
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonschema==4.22.0
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyterlab==4.2.2
jupyterlab-server==2.27.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mistune==3.0.2
nbconvert==7.16.4
nbformat==5.10.4
notebook==7.2.1
numpy===1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pickleshare==0.7.5
Pillow==10.3.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
#protobuf==5.27.2
protobuf>=3.20.3,<5.0.0

transformers==4.25.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycodestyle==2.12.0
pydeck==0.9.0b1
Pygments==2.18.0
pyparsing==3.1.2
pyrsistent==0.20.0
python-dateutil==2.9.0
# pytz==2019.3 Commenting for production deployment
# pywin32==227
pywinpty>=1.1.0
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
Send2Trash==1.8.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
terminado==0.18.1
testpath==0.6.0
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
traitlets==5.14.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
validators==0.28.3
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.11
imageio==2.35.1
Werkzeug==2.1.1 
opencv-python-headless==4.7.0.72 
tensorflow==2.16.2
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1



#Werkzeug>=2.3.7  # Latest stable version
Werkzeug==2.1.1 
PyPDF2==3.0.1
docx2txt==0.8

